% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/request.R
\name{request}
\alias{request}
\title{Request data from openelections API}
\usage{
request(state, start_date, end_date = 2012, output = "dataframe")
}
\arguments{
\item{state}{Two letter state capitalized state abbreviation.}

\item{start_date}{Start date of dataset}

\item{end_date}{End date of dataset}

\item{output}{Do you want to return a "dataframe" or a "list". The list includes additional
information about the API call.}
}
\value{
A dataframe or list
}
\description{
Request data from openelections API
}
\examples{
request("MD", 2010)
request("NY", start_date = 2002, end_date = 2007)
}

